fix: harden SSO payload validation and block nonce replay
Verify SSO signatures before decoding payloads, reject malformed or
non-UTF-8 input safely, require parsed nonce parameters, and fail closed
when SSO secrets are unset.

Share the common SSO verification, nonce consumption, and signed
redirect logic between Discourse and Librarian SSO, and add coverage for
Discourse crash/security cases plus replay rejection for both providers.
