+++ django/core/jinja2/core/librarian_sso_denied.jinja
@@ -10,13 +10,13 @@
     <h2 class='title'><i class='fas fa-lock'></i> CoMSES Librarian access required</h2>
 
     <p class='lead'>
-    You are signed in as <strong>{{ request.user.username }}</strong>, but your account is not yet a
-    member of the CoMSES Librarian pilot.
+    Your account does not currently have access to the CoMSES Librarian.
     </p>
 
     <p>
-    The Librarian is an experimental assistant for exploring the CoMSES Computational Model Library
-    and is currently limited to a small group of participants. If you would like to take part,
-    please <a href='{{ slugurl('contact') }}'>let us know</a> and we will add your account.
+    The Librarian is an <em>experimental assistant</em> for exploring the CoMSES Computational Model Library
+    and access is currently limited to a small group of pilot participants. If interested in participating,
+    please <a href='{{ slugurl('contact') }}'>send us a brief note describing how you plan to use it</a>, and
+    we will review your request.
     </p>
 {% endblock %}
